pass float variable in abs function

+++ camastersoftware.com/app/Views/firm_panel/mainpage/tx_cldr.php
@@ -275,7 +275,7 @@ if(!empty($taxCalStateAdminCookie) || !empty($taxCalFinYearAdminCookie))
                                                             $datediff = $now - $your_date;
                                                             
                                                             if($your_date>$now)
-                                                                echo abs(round($datediff / (60 * 60 * 24)));
+                                                                echo abs((float)round($datediff / (60 * 60 * 24)));
                                                             else
                                                                 echo "-";
                                                         ?>
