+++ camastersoftware.com/app/Views/super_admin/main_menu/date_wise_tax_calendar.php
@@ -256,7 +256,7 @@ if(!empty($taxCalStateCookie) || !empty($taxCalFinYearCookie))
                                                             $datediff = $now - $your_date;
                                                             
                                                             if($your_date>$now)
-                                                                echo abs(round($datediff / (60 * 60 * 24)));
+                                                                echo abs((float)round($datediff / (60 * 60 * 24)));
                                                             else
                                                                 echo "-";
                                                         ?>
